Move discovery writing to happen earlier in the pipeline (just before metrics are serialized).
This exposes a deficiency in the pipeline design: the pipeline is designed so that *Functions are changed.  When they mutate the inputs (say, to a list of booleans), then it's hard to add *Functions at the end to work on the original data.  It would be nice to support "branching" so that two *Function classes could share the same inptus.  This is almost supported via asynchronous functions, but we aren't all the way there.