Create ele-blueflood uberjar
Correctly configure ele-blueflood as a child of the parent blueflood
project, and run assembly:assembly target when using the package target.

This will make it possible to build jars for all child modules from the
parent, and also create the uberjars when using the package target. The
package target makes it possible to use another module as a dependency
without having to install those modules into your local maven
repository; the assembly:assembly target does not support this. By
piggy-backing assembly:assembly on package, we get the best of both
worlds.