fix integration tests
Integration tests were broken on systems that resolved
cassandra-maven-plugin from the lib dir. This occurred because when I
used the deploy:deploy-file target to add cassandra-maven-plugin to the
lib dir, I did give it the location of cassandra-maven-plugin's pom file
we have in the lib dir. That pom file lists dependencies for
cassandra-maven-plugin.

The fix was to re-add cassandra-maven-plugin to the lib directory
passing the location of the pom file using pomFile this time.