Further streamline use of generics wrt `Rollup`.
Rollup type is responsible for calling the right method. This replaces the `instanceof` chains in `Rollup.java`.

If we want to take this further, `SimpleNumber` could be an instance of `Rollup` whose operations are basically identity.

Then we could get the type erasure out of `AstyanaxReader.getDataToRoll()`.  (would return `Points<Rollup>`.