Make TAU processor threadpool size configurable
The default type and unit processor threadpool size of 10 was causing a
massive bottleneck during ingestion. It would hold up the rest of the
ingestion pipeline during initial load of the metadata cache and when
persisting large loads of new entries in the metrics_metadata CF, as it
does the writes one at a time.