Handle long frames by return 400.
Netty is weird in that in tries to create a message out of whatever is left in the buffer after I return the 400. Likely this is a problem of me not knowing how to do something right within netty, but I couldn't figure it out after a little searching.

We just squelch the empty message errors for now.