Remove abort rejection handler from ingest pipeline
We do not want to use the abort rejection handler on the batch writer.
This will discard writes if the incoming data starts to overwhelm the
batch writers, and we won't even be aware of it unless we look at the
logs.

We should use the default rejection handler which will still process the
data.