Fix failing rollups
Rollups are occasionally failing in prod because sometimes the var `o`
is an `Integer`, and we try to cast it to be a `Long`. Which fails
of course with a `ClassCastException` exception.

As just a couple lines above this bug `(Long) val` is created by turning
`o` into `Long` the correct way, I surmise this was a (freudian) typo.

NEEDS MORE LINT!