Do not use abort rejection handler for DiscoverWriter threadpool
Do not use abort rejection handler for DiscoverWriter threadpool. The
abort rejection handler throws an exception and causes work to be lost
if no worker is available for a job.

We do NOT want to throw away work, so use the default rejection handler
which effectively ensures all work is done, and effectively throttles
ingestion.