Add a script to check for status on compactions
We need a way to see how much longer compactions will take.  This script adds that support.  It merely prints out the number of hours remaining for any major compaction running on a node in our cluster.

Run with:

  parallel -j20 "ssh db{1}.iad.prod.bf.k1k.me 'bash -s' < ./compactionEta.sh 9080" ::: 00
01
02
03
04
05
06
07
08
09
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31 | sort

(oh, yeah, install GNU parallel (brew install parallel on a Mac) before doing it this way)