Use bounded queue for metadata batch threadpool
When `unboundedQueue` became an alias for `SynchronousQueue`, it
improved performance in asyncChain, and almost everywhere in blueflood.
Except for the metadata batch threadpool. This is one of those areas
where a large queue is required in order for work to build up. The more
work that builds up, the bigger the batches, the more efficient the meta
cache batch operations get.