Remove DropWizard for now
We haven't touched the DropWizard code in ages.  We haven't implemented it.  It's getting in the way of builds, and probably confusing people who are looking at our codebase.

Let's just remove this for the time being.  We have source control -- if we need to deliver this at some future point we'll resurrect it.