Add thread EnumValidator, which valid enum metrics and - mark metrics as bad if exceeds threshold, or - write elastic search enums index of values for querying.
This is done via a runnable thread EnumValidator that is stared by the RollupService,
for any batch of metrics that are involved in the rollup.  The EnumValidator threads
process the metrics in parallel to rollup, and does not depend on data from the rollup, except for
the locators of the metrics affected.

EnumValidator threads are manage by a threadpool executor enumValidatorExecutor, added to the RollupService.