Don't back up configuration values as 'original.*'
This backup feature makes the logic of getStringProperty confusing. I
did a case-insensitive search through all source files and couldn't
find any part of the code that used it. I even did "git log -Soriginal"
and found no related uses. It was originally added by gdusbabek in the
initial commit. I get the impression that it's for debugging purposes
only.