Make the props and defaultProps fields public.
This will aid in testing. It is only a temporary measure. I'm about to
add some tests to thoroughly test the creation of 'original.*' entries.
These tests need to be able to setup the keys and values in the props
object to a pre-defined state, and to assert against the contents of the
props object wiithout going through the getStringProperty method.