Add a couple tests to show odd behavior.
These are basically two copies of the same test intended to highlight
two counter-intuitive aspects of the behavior of getStringProperty.
Once an 'original.*' entry is created, the associated config value will
NOT be overridden by a system property. Also, once the 'original.*' is
created, it will never be updated (although this is more-or-less
expected). These behaviors must be taken into account while
refactoring, lest we break some part of the system that unwittingly
depends upon them.