Create an interface and adapter for thread pools.
Some constructors args, method args, and field are hard-coded as
ThreadPoolExecutor. This makes the threaded code difficult to test in
isolation. Normally, we would just use ExecutorService interface
instead, but RollupBatchWriter is hard-coded to use methods of
ThreadPoolExecutor. So we will use SizedExecutorService instead, and
wrap the executors in adapter objects. That way we can run code in
isolation when testing.

Eventually, RollupBatchWriter should be fixed so that it doesn't need
to call those methods. At that point, we can remove both
SizedExecutorService and the adapter class.