12th Iteration: - Resolved issue of single-point-of-failure on rest client connection. Now, code will always use all of the possible endpoints. - if one ES server dies, rest client will keep running by using other hosts. Max try to connect to any host is 10. If even after 10 iteration ES rest client couldn't connect to any host, it will simply come out.
11th Iteration:
- Brought back and refactored classes ElasticTokensIO and ElasticTokensIOTest so that they are using ES REST API and not native ES client.
- Refactored rest of the code accordingly.
- Replaced exception with logging the error, just to minimize the impact of the overall change.

10th Iteration:
- Grafana was broken as couple of classes were removed. I have added them back in this checkin.
- Made other related changes as I brought in these classes.
- Also, updated docker-entrypoint.sh and Dockerfile to get the latest image.
- Ran all of the test cases using run-test.sh, and all of them passed.

9th Iteration:
- Added refresh to the test automation rather than hard-coded sleep.
- Changed metrics mapping based on 2.4.1 settings changes
- Changed hard-coded size to the limit that ES recommends

8th Iteration:
- One of the test cases in ElasticIOIntegrationTest was failing. It's fixed now.

7th Iteration:
- Removed bunch of commented code.

6th Iteration:
- Finally all of the tests are passing.
- Bug fixes related to multiple queries.
- Fixed test class initialization and test class cleanup.
- Fixed and refactored impacted test cases because of the major change of replacing ES native client with ES REST API.

------------------------------
Bug fix in DummyDiscoveryIO4 class. Added removed constructor.
------------------------------

4th Iteration:
- Fixed a bug in ES Rest API Helper class
- Fixed many test cases, where we were not cleaning up after the test was run
- Removed a couple of test cases on dead code

3rd Iteration:
- Fixed some bugs in the code. They were related to tenantId not indexed, unit not indexed, replaced "wildcard" to "regexp" in search
- Fixed lots of test code.

This is still very much "work-in-progress"

2nd Iteration:
- Removed commented block of dead code
- Removed a couple of classes that are not in use
- Removed entry of native ES client library and related test library from the POM file
- Temporarily commented quite a few test classes to get code compiling. (NOTE: next I am working on this.)

------------------------------------------
Goal of these changes is to remove ES native client dependency, because for every ES version change, we have to make code changes.
This is much bigger cost on any upgrade. So, ES native client library is being replaced by ES REST API call.
Code changes are done for index and query for metrics and events.
Following are the major changes:

Addition of a helper class ElasticsearchRestHelper to help make this REST API calls to ES.
Removed native client use, and updated metrics query code to perform the same business logic using REST API in AbstractElasticIO and code refactor because of this.
Removed native client use, and updated metrics indexing code to perform the same business logic using REST API in ElasticIO and code refactor because of this.
Removed native client use, and updated event query and search code to perform the same business logic using REST API in EventElasticSearchIO and code refactor because of this.