fixes failing integration test
This creates a conventional 'resources' dir in each of the
integration-test directories and places config files in them to be
loaded for integration testing.

The HttpEventsQueryHandlerIntegrationTest failed when I ran 'mvn
verify'. The reason for failure was that it was expected to return
CORS headers and didn't. I traced this back to the fact that
HttpMetricsIngestionServerShutdownIntegrationTest ran first, which
caused the HttpResponder class to be initialized and read the default
CORS_ENABLED value of false. Later, when the setting was changed by
HttpIntegrationTestBase, which the failing class inherits from, it
didn't affect HttpResponder because it uses static fields and
initialization that can only happen once.

The simplest solution would be to have the shutdown IT inherit from
the common base class, but it starts and manipulates its own server,
which conflicts with those started by the base class.

This means the integration tests need to be configurable without
assuming all will inherit from a common base class.  Along the way, I
discovered that Maven and IntelliJ were treating the failsafe plugin's
'argLine' settings differently, and it was broken in both cases. The
sum of the changes in this commit can be described as migrating the
settings from both the HTTP IT base class and the argLine setting into
src/integration-test/resources, where they arguably belong anyway.

Note: I omitted the JVM settings in the argLine because MaxPermSize is
no longer a thing, and if we need to increase the heap space, we can
do it in something like .mvn/jvm.config.