consolidates and cleans up surefire configs
For the uninitiated, surefire is the plugin that runs the project's
unit tests. It doesn't run integration tests. The failsafe plugin does
that.

Currently, the surefire plugin is configured in all but one child pom,
and mostly the same. The one that's missing it is missing the ability
to skip tests by enabling the skip-unit-tests profile. Rather than
making one more copy/paste, this moves the plugin config to the parent
pom, which all children will inherit.

It omits the plugin version because that's set in pluginManagement.

It omits forkMode=always because that spawns a new JVM for each test!
This is insanely slow: ~4 minutes with it vs ~1 without.

It omits parallel execution because it seems to cause test
interference everywhere, not just in core, where it's already
disabled.  Parallel also doesn't seem to have any great impact on the
length of the test run.

It omits the kafka.config entry from the rollupTools module becaue
that module doesn't even have any unit tests.