fixes the demo environment
The `demo` folder holds some docker images and scripts aimed at
quickly starting an all-in-one blueflood server and ingesting and
querying metrics from it. It was badly broken due to age, and this
gets it up to date with good instructions for using it.

To list a few of the bigger issues fixed here:

- Updates Python scripts from 2 to 3

- Adds a container for running the Python scripts so that you don't
  have to mess with your system python

- Updates server image to Ubuntu 20.04 because there's no Cassandra
  package available for 12.04 anymore.

- Tweaks the Cassandra config for the newer version available in
  Ubuntu 20.04. I didn't go through the existing config to see if any
  of it made sense. The only changes here are the bare minimum to get
  it running.

- Updates the Cassandra schema script to use cqlsh.

- Fixes up the server "start.sh" script because some of it was out of
  date, some was unnecessary, and it wasn't waiting long enough for
  Cassandra to start, which caused everything to fail.