removes 'all-modules' profile
The only thing this profile does is add the `blueflood-all` submodule
to the build, which builds the uber jar in the package phase.  I see
no reason to have this be a special step that has to be done each time
you want to build the whole project.  This makes it a normal submodule
that's always in the build. Now a `mvn package` alone will build the
uber-jar, as you might expect from any given maven project.