updates readme content for easier start
I had some trouble working out how to get the project started for the
first time. The documentation and scripts in the project are/were
inconsistent and pointed in several different directions. Now that I
have a handle on things, I've been simplifying, removing the noise,
and improving the documentation. I've updated the wiki a lot. This
updates the basic project readme to give a good entry point for
newcomers.

Ideally, this will let new people get started with Blueflood in just a
few minutes without having to search around the wiki or read through
the source to see what's going on.