moves ES IT to the IT module
The ElasticIOIntegrationTest is in the elasticsearch module, and it's
not being run because the top-level surefire config excludes
*Integration*.java, which is the naming scheme used for this
project.

As an aside, the standard naming scheme for integration tests,
provided by the failsafe plugin, is *IT.java, so integration and unit
tests would never conflict anyway. It's always better to use
convention over configuration, but somebody here either didn't know
about it or decided to ignore it. We could consider renaming the
integration tests to follow the standard and then removing the excess
configuration from the project. In any case, this integration test
needs to move to where all the others of its kind live.

There's still a problem here, though, which is that this test can't
work without an external elasticsearch instance. If you set up a
docker ES, it works, but there's nothing in maven or the test classes
that would start an ES instance for the test. From a brief search,
there doesn't seem to be a good way to run an embedded ES for testing.
Starting a container with maven will probably be the best way forward.
For now, it's @Ignored to keep it from failing builds.

Note: these tests all pass if I start a docker ES instance and
bootstrap it with the init-es.sh script.