moves common test dependencies to top level
Junit, Powermock, Hamcrest, and Mockito are libraries that should
always be there in any module. There's no reason to repeat them,
though. Just set them in the parent pom, and the children will inherit
them. Otherwise, you'll end up with different submodules having
different combinations of the dependencies, and nobody wants to deal
with that.

Along the way, this deals with some other problems. It upgrades to the
latest junit 4 so that I can use assertThrows() in tests. It also
removes mockito-all in favor of mockito-core. The "all" variant has
hamcrest nested inside it, and they're a different version than the
other hamcrest dependencies, causing some weird internal errors.