enable the Elasticsearch integration test
This test class hasn't been running, and it's important. The problem
is that it doesn't currently have an Elasticsearch instance to run
against. This adds some options for running a test Elasticsearch and
selects one (tlrx) as the best, current option. The tlrx test library
is used elsewhere in the project for tests that need
Elasticsearch. See the comments in the test class for future-looking
options.