add a minikube env/config
Building from the existing Docker image and docker-compose
sub-projects, this adds a Kubernetes yaml file and instructions for
running it in minikube. It's very basic and needs some changes before
it could be deployed to another k8s cluster. Specifically:

1. The file mounts need to change to ConfigMaps, or else find another
way to configure the services, like the Grafana dashboards that are
available at startup here.

2. All the components are configured as simple Pods. They should be
Deployments or StatefulSets for a fault-tolerant, production workload.