make 'getting started' config more complete
As I worked with a basic blueflood started with this config, I found
some features didn't work properly because some additional modules
need to be turned on to support them.

This sets the token discovery module and associated feature flag so
that individual tokens get indexed on ingestion (tokens of com.foo.bar
are like com, com.foo, and com.foo.bar).

It also sets the events module, which enables the /events endpoint,
where not-exactly-time series data can be stored.