update base image in Dockerfile
The `java` base image is deprecated. `eclipse-temurin` is one viable
alternative for apps based on Java 8. This changes to that image. The
newer image doesn't have the old python 2 packages available, so I had
to update to python 3. Then cqlsh 4.1.1 (used in the container startup
to initialize the database) didn't work, so I updated it to a newer
version which should work all the way up to cassandra 4, at least.