improve docker entrypoint script
A few small changes to make the script better.

1. Cassandra may be listening on port 9160 for thrift (old style) or
9042 for cqlsh (new style) connections.

2. If ROLLUP_KEYSPACE wasn't set, it would make the cqlsh init script
invalid by replacing all occurrences of the keyspace "DATA" with "".

3. Assuming java is installed at /usr/bin/java makes the image
brittle. Let it search the PATH, the way it's designed to.

4. For dev and debugging purposes, it's nice to be able to attach a
remote debugger to blueflood when it's running. A couple of env vars
added to the script let users enable remote JVM debugging on demand.