make init-es.sh work outside bash
Many Docker images are built on lightweight linux distros that don't
include bash. The eclipse-temurin image that provides a JDK is one
example. To run a script like this on that kind of image, it can't use
any exclusive bash-isms. The script just has one, small issue: it uses
the `BASH_SOURCE` env variable. An almost identical replacement for
`${BASH_SOURCE[0]}` is `$0`, and it should work fine for us. For
further details see https://stackoverflow.com/a/35006505.