unsynchronize TokenCache
Like the LocatorCache before, TokenCache is unnecessarily
synchronized. It's built on a concurrency-safe guava cache, so
synchronizing will only slow things down. It turns out the TokenCache
isn't used quite as heavily as the LocatorCache because "leaf" tokens
aren't cached in the TokenCache, but production usage is still
significant.

This adds testing of the TokenCache (missing previously), including
some concurrency testing to show it's safe to unsynchronize. The
concurrency tests borrow heavily from LocatorCacheTest.