resolve tempus fujit snapshot version
I originally used 1.2-SNAPSHOT because it had a feature I thought I
wanted but ended up not using. I think it was running tests in
parallel.  I ended up writing my own concurrent test runner because I
needed it to work a certain way for the LocatorCache and TokenCache
tests.