Added github action to trigger when PR is created
We need to create an automated check which tests for unit and integration test cases whenever a new pull request is created. This is required because automated PR checks helps in detecting any failures which new changes might bring in.
So, We are using github action `on pull_request` trigger and maven to perform tests and generate reports.