add support for multiple modules to loader
Module loading is a mixed bag at the moment. Some parts of Blueflood
allow for multiple modules of the same type. See for example
DiscoveryWriter's use of DISCOVERY_MODULES. Other parts require at
most one module, such as HttpMetricsIndexHandler's use of the same
DISCOVERY_MODULES.

Given that the names of the module properties are plural and that
logic exists for parsing multiple modules from one config item, like
in DiscoveryWriter, I think the original design was for zero to many
modules of each type to be configurable so you can plug in multiple
implementations of something like Discovery.

To that end, this makes the ModuleLoader support multiple modules of
the same type. It keeps them separate using a new concept called a
"qualifier". This is a shameless ripoff of Spring's qualifier concept,
where you can configure multiple beans of the same type and use a
qualifier to autowired the right one in different places.

This change doesn't address the need to get all modules of a given
type, regardless of qualifier. Since DiscoveryWriter already handles
that on its own by going behind ModuleLoader's back, I'll leave that
for a future change. It goes without saying, though, that ideally all
module loading would go through the ModuleLoader.