remove excess token discovery code
When someone added token indexing alongside the name indexing, they
didn't do a good job integrating it with the existing code. They
basically duplicated DiscoveryWriter and all the things it does. It's
much more efficient and logical for token discovery to live behind the
DiscoveryWriter in a new module alongside the module that provides
name discovery. DiscoveryWriter already knows how to load multiple
discovery modules and use them. This also lets all discovery
activities fall under the DiscoveryWriter and any features it
provides, like caching and throttling to reduce load on discovery
backends. As it stands now, any such improvments would have to be
duplicated in TokenDiscoveryWriter.

The main goal of this change is to remove the duplicate Token-related
classes/interfaces and move the ElasticTokensIO, which is the real
logic of the token indexing, into the DiscoveryWriter. To keep the
change as minimal as possible, I haven't reworked all the tests that
use the existing ElasticTokensIO contract, so I had to add a hacky
`insertDiscovery(List)` method there to handle either a list of
metrics or a list of tokens. Only tests will pass a list of tokens
there. At runtime, it's always a list of metrics, passed by the
DiscoveryWriter. I'll clean this up in a following change.

To implement the change, a small bit of code that manages the token
cache has to move down from TokenDiscoveryWriter to ElasticTokensIO,
and then ElasticTokensIO changes to a normal DiscoveryIO that's
plugged into the DiscoveryWriter.