Added manual GitHub action trigger to create release package.
We need a trigger which can be used to create an automated release package for blueflood. We are using GitHub action with maven release plugin here to create the release package.
Maven release package helps in performing sequence of automated steps which will run tests on latest master code and then will create the tag which will have the release version which we specified when triggering the workflow. It will also increment the next snapshot version in all the moms wherever required.
A PR will be created as part of this workflow which will have the pom version changes of incremented snapshot version for new development.
Artifacts will be copied to the release package as well.