move dependency retrieval to constructor
From history, it seems choosing the handler was at one point dependent
on details of the request. That reature (enum metrics iirc) was
removed from the project, but this line was left in the middle of the
request processing, when it really belongs in the constructor. There's
no reason to look up the handler for every request.