let env vars set config properties
As it stands, configuration items have default values which can be
overriden with a config file, and both can be overriden by setting
system properties. For containerization, it'll be important to also
let environment variables be used. This adds environment variables as
the highest-precedence configuration override, so we now have four
levels to find a config item. In precedence order, they are:
environment variables, system properties, config file, defaults.