let Docker container start without initting
For a simple dev or demo environment, it's convenient to have the
Blueflood container automatically initialize Cassandra and
Elasticsearch for you when you start. The Docker image does that
today. For anything like a realistic deployment, you'd never want this
to happen automatically. This adds two env vars that will stop the
initialization from running. It also cleans up the init code
slightly. Specifically, it adds a 1-second timeout to the netcat
probes that wait for Cassandra and Elasticsearch to come up, and it
watches for Elasticsearch on either port 9200 (rest port) or 9300
(native transport port).