allow docker image to take config files
Through working with the image and deploying to Kubernetes, it became
apparent that this Docker image was made with limited, dev-style
deployments in mind. In particular, we don't want to have to configure
everything via environment variables. We want to have a config file
that holds most of our Blueflood settings and override a few things
with the environment.

This makes the Docker image accept an existing file location for the
Blueflood config and the log4j config. If either is missing, it falls
back to its original behavior of: 1) generating a Blueflood config
file by dumping all environment variables to a file and 2) writing a
static log4j config file where only the log level of the Blueflood
classes is configurable by the user.