clean up Docker image env
Formerly, the Docker image was built to dump a lot of Blueflood config
into the env and then dump the whole env into a blueflood.properties
file to configure the Blueflood process. Now the image can be given a
specific config file, and env vars serve as the highest-precedence way
of overriding blueflood settings. Given that most of these are just
repeating the default settings anyway, there's no reason to keep them
around.

The settings for heap space are the only two that are important
because the docker command for starting Blueflood relies on them. This
makes them configurable in the entrypoint script so that they have
reasonable defaults but can be set via env var from the docker command
line or k8s config, for example.

btw, `-Xms` is initial heap size, not "min heap size". I'd rename that
variable, but I don't want to affect anyone/anything that might be
using it.