ES test server is global
Starting a new Elasticsearch for every test that needs it is expensive
in terms of time. This creates a global singleton for managing ES test
servers so that ideally, every integration test can share the same
server.

We still need to migrate all existing IT's to use this class so that
we can easily transition to using TestContainers eventually (read:
very soon).