externalize Elasticsearch IT config
The new ElasticsearchTestServer has the capabilities needed to start
testing against new versions of Elasticsearch. This adds configuration
settings for choosing how the test suite will start
Elasticsearch. Right now, it makes it easy to run the tests manually
against whatever you want. In the future, it could be used in the
build pipeline to run a test matrix against two or more Elasticsearch
versions when preparing for upgrades.