recategorize two tests as integration tests
These tests are running as unit tests and starting an Elasticsearch
server. To make the best use of ElasticsearchTestServer, we want all
tests that use Elasticsearch to run in the same set of tests, so that
we only start Elasticsearch once. This moves/renames those two tests
so they run with all the other integration tests.