migrate all ES tests to ES test server
This removes all uses of the Elasticsearch tlrx test library except
for the one contained in ElasticsearchTestServer. Then everything that
was using that library uses the new test class instead. This is so
that we can easily switch to using test containers and also centrally
manage the Elasticsearch version in tests.

I also cleaned up all of the teardown code I could find where an HTTP
client was being used to delete data from Elasticsearch. In some
cases, it was implemented wrong and was deleting entire indexes
instead of just deleting documents out of the indexes. This was the
cause of some of our spurious test failures. When the index was
deleted, its mapping was lost, and the next test that used it created
an index with a dynamic mapping.

The new reset code is safe and guarantees correct indexes for every
test. It's run at the start of each test class either by the class
itself, or because the class extends HttpIntegrationTestBase, which
sets everything up for you.