prevent mbean re-registration
Is this the right thing to do? In tests, these classes get
instantiated multiple times, and it causes errors when they try to
register themselves again. We can avoid that by checking if they're
already registered.  It means we lose the notification from the logs
if something were being mishandled in a real, runtime environment.