de-static-ify DatastaxIO
Much of Blueflood's core code is built as static singletons, but
DatastaxIO is just pure static. It connects to the database at class
initialization time, making it impossible to touch this thing without
having a database up and running. I want to add the ability to
initialize the database schema from code, but I can't do that because
this requires the keyspace to already exist just to load the class!

The only important method here is `getSession()`, which all other code
uses to get access to the database. As long as that method's contract
is preserved, we can make changes to the rest of the class around it.

This changes DatastaxIO to be a static singleton in line with the rest
of Blueflood and leaves the constructor public so that other code can
get access to the database when necessary. There are no unit tests of
this class, but it's the central access point for all database traffic
that uses the Datastax driver, so integration tests will certainly
fail if there's a problem here.