better elasticsearch failures
When Elasticsearch returns a non-2xx response code,
ElasticsearchRestHelper throws an exception, but the reason for the
error is lost. This adds the response body to the message. The body
almost always has useful information in it, especially when the
problem is because you're making some bad request or other while
modifying the code.