elasticsearch fixture can reset to clean state
Many of the integration tests expect to be running with pristine
Elasticsearch indexes. Otherwise, they see conflicting data from other
tests and fail. There are better ways to deal with this. For example,
use a random tenant for each test. For now, being able to reset our
test Elasticsearch to a clean state is a surefire way to cover our
bases, even though it'll take extra time.

This implements a reset() method that simply deletes all indexes with
an http client and then runs the init again. Since that means we'll be
using the init script on the tlrx test instance, I went ahead and
changed the init for that to use the script, too. Now everything gets
initialized the same way!

As a proof of concept, I've made ElasticIOIntegrationTest use the
method instead of the manual cleanup it was doing before.