update ES 6 readme
Adds more discoveries of Elasticsearch 6 behavior to the README for
that version of init files. In particular, I figured out what the
custom analyzer in the old `index-settings.json` does and that while
it could work in version 6, it's probably better not to do it that way
and use the new tokens code instead.