add missing index to integration test
I discovered why ElasticIOIntegrationTest tends to be flaky. In its
setup, it writes metrics do a new/different Elasticsearch index, and
it configures ElasticIO to search both the standard index and the new
one. I'm not quite sure why that is, unless it was an early attempt to
ensure the code would work with time-based rolling indexes. Whatever
the case, the new index used to be initialized with a correct mapping,
but that was lost at some point, so now the mapping is generated
dynamically, so the data doesn't get indexed correctly, leading to
bizarre-seeming test results.

This adds a method to the test server to allow for creating this index
with an appropriate mapping so that the data is indexed correctly, and
the test is testing what it was originally meant to.