wait for ES to catch up in indexing test
This test seems to execute too quickly for Elasticsearch to be ready
with the new data. It used to have a "refresh index" call in it, but I
removed it because it used the old tlrx library. According to
https://www.elastic.co/guide/en/elasticsearch/reference/current/indices-refresh.html,
Elasticsearch refreshes indexes every second by default. This must be
a case where the time from test setup to execution is under a second,
and it's regularly returning an empty result. If we wait a little bit,
everything is fine. It seems our options are to wait, manually
refresh, or set a lower refresh interval.

I've added an await to fix it here. If this problem pops up in other
tests, we can consider the other options.