Changes for log4j2 upgrade
Here we make all the changes need for log42 to work. Some of the changes are in pom file to update log4j2 version and support it. We need to remove some code as they are not compatible in log4j2. For PropertyCOnfigurator change there is a property monitorInterval in log4j2. We also need to update the Metrics version to match log4j2 support.